requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alibabacloud-tea==0.3.9
alibabacloud-waf-openapi20190910==1.2.1
alibabacloud-waf-openapi20211001==4.6.0
alibabacloud_credentials==0.3.5
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.2
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.11
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
aliyun-python-sdk-alidns==3.0.7
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-waf-openapi==1.1.10
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
dnspython==2.6.1
frozenlist==1.4.1
httplib2==0.22.0
idna==3.8
jmespath==0.10.0
multidict==6.0.5
pycparser==2.22
pyparsing==3.1.4
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4
